insurance-claim-agent/requirements.txt
fastapi
uvicorn[standard]
pydantic
pymupdf
pytesseract
pillow
scikit-learn
numpy
python-multipart
streamlit
pyyaml
easyocr
google-genai
python-dotenv
